pyproject.toml
[project]
name = "excel-gigachat-analyzer"
version = "0.1.0"
description = "Console application to analyze Excel data using GigaChat API"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "pandas>=2.0.0",
    "openpyxl>=3.1.0",
    "python-dotenv>=1.0.0",
    "gigachat>=0.1.0",
]

[project.scripts]
excel-analyzer = "main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
]
